Crate sidechain_domain

Types/Structs/Functions meant to be used by other modules / our Business Logic (i.e the Sidechain’s node / runtime in our case)

Modules

byte_string

crypto

mainchain_epoch

Structs

AdaBasedStaking

AdaBasedStaking is a variant of Plutus type StakeOwnership. The other variant, TokenBasedStaking, is not supported

AssetId

AssetName

AuraPublicKey

BTreeMap

An ordered map based on a B-Tree.

CandidateRegistration

CandidateRegistrations

Information about an Authority Candidate’s Registrations at some block.

CrossChainKeyHash

CrossChainPublicKey

CrossChainSignature

DParameter

DelegatorStakeAmount

EpochNonce

GrandpaPublicKey

MainchainAddress

Wraps UTF-8 bytes of Mainchain Address in bech32 format. Example: utf-8 bytes of “addr_test1wz5qc7fk2pat0058w4zwvkw35ytptej3nuc3je2kgtan5dq3rt4sc” are “0x616464725f7465737431777a35716337666b327061743030353877347a77766b77333579747074656a336e7563336a65326b6774616e356471337274347363”

MainchainBlock

MainchainKeyHash

blake2b_224 hash of a Cardano Verification (Public) Key. It can be a hash of Payment Verification, Payment Extended Verification, Stake Pool Verification Key or Staking Verification Key.

MainchainSignature

EDDSA signature

McBlockHash

McBlockNumber

A main chain block number. In range [0, 2^31-1].

McEpochNumber

A main chain epoch number. In range [0, 2^31-1].

McSlotNumber

McTxHash

McTxIndexInBlock

An index of transaction in a block. In range [0, 2^31-1].

NativeTokenAmount

PermissionedCandidateData

PolicyId

Cardano Policy Id

PoolDelegation

RegistrationData

UTxO Output of a Registration Transaction on Cardano

ScEpochNumber

ScSlotNumber

SidechainPublicKey

SidechainSignature

StakeDelegation

Amount of Lovelace (which is a fraction of 1 ADA) staked/locked on Cardano

StakeDistribution

StakeKeySignature

EDDSA signature made with Stake Signing Key

StakePoolPublicKey

StakePublicKey

TransactionCbor

CBOR bytes of Cardano Transaction.

UtxoId

Identifies UTxO by transaction hash, and the index of output.

UtxoIndex

An index of output of a transaction. In range [0, 2^15-1].

UtxoInfo

UtxoInfoOrderingKey

VKeyWitnessCbor

CBOR bytes of Cardano VKeyWitness.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Enums

DelegatorKey

NetworkType

Constants

MAINCHAIN_SIGNATURE_LEN

EDDSA signature, 64 bytes.

MAX_ASSET_NAME_LEN

STAKE_KEY_SIGNATURE_LEN

EDDSA signature, 64 bytes.

TX_HASH_SIZE

Functions

offset_data_epoch

Type Aliases

ScriptHash