Crate sidechain_domain

Types/Structs/Functions meant to be used by other modules / our Business Logic (i.e the Sidechain’s node / runtime in our case)

Modules

byte_string

crypto

mainchain_epoch

Structs

ATMSPlainAggregatePubKey

Hash of concatenated private keys. It should match: – | Invariant: ‘ATMSPlainAggregatePubKey’ is the sorted concatenated hash of – sidechain public keys. More precisely, – @ – committeePubKeys = sort([key1, key2, …, keyN]) – committeePubKeysHash = blake2b_256(concat(committeePubKeys)) – keyi - 33 bytes compressed ECDSA public key of a committee member – @ newtype ATMSPlainAggregatePubKey = ATMSPlainAggregatePubKey ByteString https://github.com/input-output-hk/partner-chains-smart-contracts/blob/5b19d25a95c3ab49ae0e4c6ce0ec3376f13b3766/docs/Specification.md#L554-L561

AdaBasedStaking

AdaBasedStaking is a variant of Plutus type StakeOwnership. The other variant, TokenBasedStaking, is not supported

AssetId

AssetName

AuraPublicKey

CandidateRegistration

CandidateRegistrations

Information about an Authority Candidate’s Registrations at some block.

CommitteeHash

CrossChainPublicKey

CrossChainSignature

DParameter

EpochNonce

GrandpaPublicKey

MainchainAddress

Wraps UTF-8 bytes of Mainchain Address in bech32 format. Example: utf-8 bytes of “addr_test1wz5qc7fk2pat0058w4zwvkw35ytptej3nuc3je2kgtan5dq3rt4sc” are “0x616464725f7465737431777a35716337666b327061743030353877347a77766b77333579747074656a336e7563336a65326b6774616e356471337274347363”

MainchainAddressHash

Some hash of MainchainAddress, 28 bytes. Presumably blake2b_224. Way to get it: cardano-cli latest address key-hash –payment-verification-key-file FILE

MainchainBlock

MainchainPrivateKey

MainchainPublicKey

MainchainSignature

McBlockHash

McBlockNumber

A main chain block number. In range [0, 2^31-1].

McEpochNumber

A main chain epoch number. In range [0, 2^31-1].

McSlotNumber

McTxHash

McTxIndexInBlock

An index of transaction in a block. In range [0, 2^31-1].

NativeTokenAmount

PermissionedCandidateData

PolicyId

Cardano Policy Id

RegistrationData

UTxO Output of a Registration Transaction on Cardano

ScEpochNumber

ScSlotNumber

SidechainPublicKey

SidechainPublicKeysSorted

SidechainSignature

StakeDelegation

Amount of Lovelace (which is a fraction of 1 ADA) staked/locked on Cardano

UtxoId

Identifies UTxO by transaction hash, and the index of output.

UtxoIndex

An index of output of a transaction. In range [0, 2^15-1].

UtxoInfo

UtxoInfoOrderingKey

ValidatorHash

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Enums

NetworkType

Constants

MAINCHAIN_ADDRESS_HASH_LEN

MAX_ASSET_NAME_LEN

TX_HASH_SIZE

VALIDATOR_HASH_LEN

Type Aliases

ScriptHash